requirements.txt
pyModbus
pybase64
importlib-metadata
requests
Python-Deprecated